Updated UI and added Electrostats toggle + Randomizer
* Moved UI into 3 separate files
* Moved mod settings to Mod Buttons (in Main Menu) instead of Mod Settings
* Added additional stuff to the config to save if Electrostats toggle is on etc.
* Added about section that indicates what was done by who and redirects to the respective user's GitHub
* Started adding Profile changer to be added as a tab in-game to allow for changing profiles while in TournamentAssistant, Multiplayer or similar. This feature is intended to more easily be able to use different profiles on different maps.
