+++ HarmonyPatches/SaberTailorPatches.cs
@@ -32,7 +32,7 @@ namespace SaberTailor.HarmonyPatches
         {
             if (instance != null && IsPatched)
             {
-                instance.UnpatchAll(InstanceId);
+                instance.UnpatchSelf();
                 IsPatched = false;
             }
         }
