+++ SaberTailor.csproj
@@ -9,6 +9,7 @@
     <RootNamespace>SaberTailor</RootNamespace>
     <AssemblyName>SaberTailor</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
+    <LangVersion>8</LangVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <TargetFrameworkProfile />
@@ -38,9 +39,16 @@
     <Compile Include="HarmonyPatches\Patches\SaberTrailRendererOnEnable.cs" />
     <Compile Include="HarmonyPatches\SaberTailorPatches.cs" />
     <Compile Include="HarmonyPatches\Patches\AdjustPlatformSpecificControllerTransformPatch.cs" />
+    <Compile Include="Installers\MenuInstaller.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Properties\Resources.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
+    <Compile Include="Settings\Classes\BaseConfig.cs" />
     <Compile Include="Settings\Classes\GripConfig.cs" />
     <Compile Include="Settings\Classes\GripRawConfig.cs" />
     <Compile Include="Settings\Classes\Int3.cs" />
@@ -49,13 +57,19 @@
     <Compile Include="Settings\Classes\ScaleRawConfig.cs" />
     <Compile Include="Settings\Classes\TrailConfig.cs" />
     <Compile Include="Settings\Configuration.cs" />
-    <Compile Include="Settings\UI\MainSettings.cs" />
     <Compile Include="Settings\Utilities\ConfigurationImporter.cs" />
     <Compile Include="Settings\Utilities\FileHandler.cs" />
     <Compile Include="Settings\Utilities\GameSettingsTransfer.cs" />
     <Compile Include="Settings\Utilities\PluginConfig.cs" />
     <Compile Include="Settings\Utilities\ProfileManager.cs" />
     <Compile Include="Tweaks\SaberLength.cs" />
+    <Compile Include="UI\LeftController.cs" />
+    <Compile Include="UI\LeftSideProfileMenuController.cs" />
+    <Compile Include="UI\TabButtonUI.cs" />
+    <Compile Include="UI\MiddleSettingsPaneController.cs" />
+    <Compile Include="UI\RightController.cs" />
+    <Compile Include="UI\UIFlowCoordinator.cs" />
+    <Compile Include="Utilities\SaberUtils.cs" />
     <Compile Include="Utilities\Utils.cs" />
     <Compile Include="Utilities\Vector3Extensions.cs" />
   </ItemGroup>
@@ -76,7 +90,7 @@
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\GameplayCore.dll</HintPath>
       <Private>False</Private>
     </Reference>
-        <Reference Include="Hive.Versioning">
+    <Reference Include="Hive.Versioning">
       <HintPath>$(GameDirPath)Libs\Hive.Versioning.dll</HintPath>
       <Private>False</Private>
     </Reference>
@@ -84,6 +98,7 @@
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\HMLib.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="HMUI, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null" />
     <Reference Include="IPA.Loader">
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\IPA.Loader.dll</HintPath>
       <Private>False</Private>
@@ -101,6 +116,8 @@
       <Private>False</Private>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Data, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL" />
+    <Reference Include="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL" />
     <Reference Include="Unity.TextMeshPro">
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\Unity.TextMeshPro.dll</HintPath>
       <Private>False</Private>
@@ -109,10 +126,13 @@
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\UnityEngine.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="UnityEngine.CoreModule">
-      <HintPath>$(GameDirPath)Beat Saber_Data\Managed\UnityEngine.CoreModule.dll</HintPath>
-      <Private>False</Private>
+    <Reference Include="UnityEngine.CoreModule, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL" />
+    <Reference Include="UnityEngine.ImageConversionModule, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL" />
+    <Reference Include="UnityEngine.UI, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>D:\BSLegacy1191\Beat Saber\Beat Saber_Data\Managed\UnityEngine.UI.dll</HintPath>
     </Reference>
+    <Reference Include="UnityEngine.UIModule, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL" />
     <Reference Include="UnityEngine.VRModule">
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\UnityEngine.VRModule.dll</HintPath>
       <Private>False</Private>
@@ -121,17 +141,35 @@
       <HintPath>$(GameDirPath)Beat Saber_Data\Managed\UnityEngine.XRModule.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="Zenject, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL" />
+    <Reference Include="Zenject-usage, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="manifest.json" />
+    <EmbeddedResource Include="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
     <EmbeddedResource Include="Resources\description.md" />
     <EmbeddedResource Include="Resources\icon.png" />
-    <EmbeddedResource Include="Settings\UI\Views\mainsettings.bsml">
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="UI\Views\LeftControllerSettings.bsml" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="UI\Views\RightControllerSettings.bsml" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="UI\Views\MiddleMainSettings.bsml" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="UI\Views\LeftSideProfileMenu.bsml" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\arrow.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>copy ".\SaberTailor.dll" "$(GameDirPath)Plugins"</PostBuildEvent>
+    <PostBuildEvent>copy ".\SaberTailor.dll" "D:\BSLegacy1191\Beat Saber\Plugins"</PostBuildEvent>
   </PropertyGroup>
 </Project>
