+++ Settings/Classes/BaseConfig.cs
@@ -1,0 +1,11 @@
+﻿using SaberTailor.Settings.Classes;
+
+namespace SaberTailor.Settings.Classes
+{
+    public class BaseConfig
+    {
+        public bool Electrostats = false;
+        public bool Randomize = false;
+        public bool electroball = false;
+    }
+}
