+++ Settings/Utilities/GameSettingsTransfer.cs
@@ -23,76 +23,83 @@ namespace SaberTailor.Settings.Utilities
 
         internal static bool ImportFromGame()
         {
-            // Get reference to MainSettingsModelSO
-            MainSettingsModelSO mainSettings = Resources.FindObjectsOfTypeAll<MainSettingsModelSO>().FirstOrDefault();
-            if (mainSettings == null)
+            try
             {
-                Logger.log.Error("ImportFromGame: Unable to get a handle on MainSettingsModelSO. Exiting...");
-                return false;
+                // Get reference to MainSettingsModelSO
+                MainSettingsModelSO mainSettings = Resources.FindObjectsOfTypeAll<MainSettingsModelSO>().FirstOrDefault();
+                if (mainSettings == null)
+                {
+                    Logger.log.Error("ImportFromGame: Unable to get a handle on MainSettingsModelSO. Exiting...");
+                    return false;
+                }
+
+                // Check for position/rotation settings
+                if (mainSettings.controllerPosition == null)
+                {
+                    Logger.log.Error("ImportFromGame: Settings for controller position not found. Exiting...");
+                    return false;
+                }
+
+                if (mainSettings.controllerRotation == null)
+                {
+                    Logger.log.Error("ImportFromGame: Settings for controller rotation not found. Exiting...");
+                    return false;
+                }
+
+                Vector3SO gameCtrlPosSO = mainSettings.controllerPosition;
+                Vector3SO gameCtrlRotSO = mainSettings.controllerRotation;
+
+                Vector3 ctrlPos = new Vector3(gameCtrlPosSO.value.x, gameCtrlPosSO.value.y, gameCtrlPosSO.value.z);
+                Vector3 ctrlRot = new Vector3(gameCtrlRotSO.value.x, gameCtrlRotSO.value.y, gameCtrlRotSO.value.z);
+
+                Configuration.GripCfg.PosLeft = new Int3()
+                {
+                    x = -(int)Math.Round(ctrlPos.x * 1000, MidpointRounding.AwayFromZero),
+                    y = (int)Math.Round(ctrlPos.y * 1000, MidpointRounding.AwayFromZero),
+                    z = (int)Math.Round(ctrlPos.z * 1000, MidpointRounding.AwayFromZero)
+                };
+
+                Configuration.GripCfg.RotLeft = new Int3()
+                {
+                    x = (int)Math.Round(ctrlRot.x, MidpointRounding.AwayFromZero),
+                    y = -(int)Math.Round(ctrlRot.y, MidpointRounding.AwayFromZero),
+                    z = -(int)Math.Round(ctrlRot.z, MidpointRounding.AwayFromZero)
+                };
+
+                Configuration.GripCfg.PosRight = new Int3()
+                {
+                    x = (int)Math.Round(ctrlPos.x * 1000, MidpointRounding.AwayFromZero),
+                    y = (int)Math.Round(ctrlPos.y * 1000, MidpointRounding.AwayFromZero),
+                    z = (int)Math.Round(ctrlPos.z * 1000, MidpointRounding.AwayFromZero)
+                };
+
+                Configuration.GripCfg.RotRight = new Int3()
+                {
+                    x = (int)Math.Round(ctrlRot.x, MidpointRounding.AwayFromZero),
+                    y = (int)Math.Round(ctrlRot.y, MidpointRounding.AwayFromZero),
+                    z = (int)Math.Round(ctrlRot.z, MidpointRounding.AwayFromZero)
+                };
+
+                Configuration.GripCfg.OffsetLeft = new Int3()
+                {
+                    x = 0,
+                    y = 0,
+                    z = 0
+                };
+
+                Configuration.GripCfg.OffsetRight = new Int3()
+                {
+                    x = 0,
+                    y = 0,
+                    z = 0
+                };
+
+                Configuration.Grip.UseBaseGameAdjustmentMode = true;
             }
-
-            // Check for position/rotation settings
-            if (mainSettings.controllerPosition == null)
+            catch (Exception e)
             {
-                Logger.log.Error("ImportFromGame: Settings for controller position not found. Exiting...");
-                return false;
+                Logger.log.Error(e);
             }
-
-            if (mainSettings.controllerRotation == null)
-            {
-                Logger.log.Error("ImportFromGame: Settings for controller rotation not found. Exiting...");
-                return false;
-            }
-
-            Vector3SO gameCtrlPosSO = mainSettings.controllerPosition;
-            Vector3SO gameCtrlRotSO = mainSettings.controllerRotation;
-
-            Vector3 ctrlPos = new Vector3(gameCtrlPosSO.value.x, gameCtrlPosSO.value.y, gameCtrlPosSO.value.z);
-            Vector3 ctrlRot = new Vector3(gameCtrlRotSO.value.x, gameCtrlRotSO.value.y, gameCtrlRotSO.value.z);
-
-            Configuration.GripCfg.PosLeft = new Int3()
-            {
-                x = -(int)Math.Round(ctrlPos.x * 1000, MidpointRounding.AwayFromZero),
-                y = (int)Math.Round(ctrlPos.y * 1000, MidpointRounding.AwayFromZero),
-                z = (int)Math.Round(ctrlPos.z * 1000, MidpointRounding.AwayFromZero)
-            };
-
-            Configuration.GripCfg.RotLeft = new Int3()
-            {
-                x = (int)Math.Round(ctrlRot.x, MidpointRounding.AwayFromZero),
-                y = -(int)Math.Round(ctrlRot.y, MidpointRounding.AwayFromZero),
-                z = (int)Math.Round(ctrlRot.z, MidpointRounding.AwayFromZero)
-            };
-
-            Configuration.GripCfg.PosRight = new Int3()
-            {
-                x = (int)Math.Round(ctrlPos.x * 1000, MidpointRounding.AwayFromZero),
-                y = (int)Math.Round(ctrlPos.y * 1000, MidpointRounding.AwayFromZero),
-                z = (int)Math.Round(ctrlPos.z * 1000, MidpointRounding.AwayFromZero)
-            };
-
-            Configuration.GripCfg.RotRight = new Int3()
-            {
-                x = (int)Math.Round(ctrlRot.x, MidpointRounding.AwayFromZero),
-                y = (int)Math.Round(ctrlRot.y, MidpointRounding.AwayFromZero),
-                z = (int)Math.Round(ctrlRot.z, MidpointRounding.AwayFromZero)
-            };
-
-            Configuration.GripCfg.OffsetLeft = new Int3()
-            {
-                x = 0,
-                y = 0,
-                z = 0
-            };
-
-            Configuration.GripCfg.OffsetRight = new Int3()
-            {
-                x = 0,
-                y = 0,
-                z = 0
-            };
-
-            Configuration.Grip.UseBaseGameAdjustmentMode = true;
 
             return true;
         }
