+++ UI/LeftSideProfileMenuController.cs
@@ -1,0 +1,18 @@
+﻿using BeatSaberMarkupLanguage.Attributes;
+using BeatSaberMarkupLanguage.ViewControllers;
+using SaberTailor.Settings.Utilities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using TMPro;
+
+namespace SaberTailor.UI
+{
+    [ViewDefinition("SaberTailor.UI.Views.LeftSideProfileMenu.bsml")]
+    public class LeftSideProfileMenuController : BSMLAutomaticViewController
+    {
+        //TODO: IMPLEMENT THIS.....
+    }
+}
