+++ manifest.json
@@ -5,11 +5,11 @@
     "#![SaberTailor.Resources.description.md]",
     "Customize stuff about your sabers"
   ],
-  "gameVersion": "1.17.0",
+  "gameVersion": "1.19.1",
   "icon": "SaberTailor.Resources.icon.png",
   "id": "SaberTailor",
   "name": "SaberTailor",
-  "version": "3.5.0",
+  "version": "4.0.0",
   "dependsOn": {
     "BSIPA": "^4.2.0"
   },
